requirements.txt
termux-api==1.2.3
inotify-simple==1.3.5
crontab==0.23.0
cron-descriptor==1.2.32
PyYAML==6.0